ir_project/requirements.txt
# Information Retrieval System - Requirements
# ==========================================

# Core Dependencies
langchain-community>=0.0.13
langchain-ollama>=0.1.0
langchain>=0.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0

# Web Framework (for API demo)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# UI Framework (for Streamlit demo)
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced functionality
python-multipart>=0.0.6  # For FastAPI file uploads
aiofiles>=23.0.0         # For async file operations
python-dotenv>=1.0.0     # For environment configuration

# Note: You'll also need Ollama installed separately
# Install Ollama from: https://ollama.ai/
# Then pull the embedding model: ollama pull nomic-embed-text:latest